Export `ConfigFile` struct which can be de/serialized from and to a configuration file (#607)
* clippy

* Export ConfigFile from api module

* Improve doc string

* Add serde defaults to ConfigFile fields

* fmt & clippy

* Update CHANGELOG
