+++ aquadoggo/src/api/mod.rs
@@ -2,9 +2,11 @@
 
 #[allow(clippy::module_inception)]
 mod api;
+mod config_file;
 mod lock_file;
 mod migration;
 
 pub use api::NodeInterface;
+pub use config_file::ConfigFile;
 pub use lock_file::LockFile;
 pub use migration::migrate;
